Entity: float to int + type box bug fixes
I've noticed that the float doesn't even work and acts like an int, so I reverted everything to being an int instead of a float

+++ source/include/Global.hpp
@@ -14,9 +14,9 @@ namespace Global
     {
         std::string name;
         
-        float x;
-        float y;
-        float direction;
+        int x;
+        int y;
+        int direction;
 
         uint16_t size;
